requirements.txt
asgiref==3.10.0
beautifulsoup4==4.14.2
certifi==2025.10.5
charset-normalizer==3.4.4
contourpy==1.3.3
cycler==0.12.1
Django==5.2.7
django-environ==0.12.0
djangorestframework==3.16.1
fonttools==4.61.0
gunicorn==23.0.0
idna==3.11
joblib==1.5.2
kiwisolver==1.4.9
matplotlib==3.10.7
mysqlclient==2.2.7
narwhals==2.14.0
numpy==2.3.5
packaging==25.0
paho-mqtt==2.1.0
pandas==2.3.3
pillow==12.0.0
plotly==6.5.0
pymongo==3.12.3
PyMySQL==1.1.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
scikit-learn==1.7.2
scipy==1.16.3
six==1.17.0
soupsieve==2.8
sqlparse==0.5.3
threadpoolctl==3.6.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0